requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary>=2.9.11
python-dotenv==1.0.0
openai>=1.50.0
httpx>=0.27.0,<0.28.0
APScheduler==3.10.4
bcrypt==4.1.2
gunicorn==21.2.0
Werkzeug==3.0.1
